+++ perf/benchmark.jl
@@ -11,7 +11,7 @@ for nthreads in (1, 4)
             println("Iteration ", j, " (", n, " datapoints)")
             t = collect(linspace(0.01, 10, n))
             s = sin.(t) .+ 1.5*cospi.(4t) .+ 3
-            plan = LombScargle.plan(t, s, fast = true)
+            plan = LombScargle.plan(t, s, fast = true, flags = FFTW.MEASURE)
             time = minimum(@benchmark(lombscargle($plan)).times)/1e9
             println(" Julia:  ", time, " seconds")
             println(file, n, "\t", time)